reusing conftest.py and logging.ini with one issue... logs no longer appear in test results console
