--- qa-pytest-examples/logging.ini
@@ -1,53 +1,0 @@
-; SPDX-FileCopyrightText: 2025 Adrian Herscu
-;
-; SPDX-License-Identifier: Apache-2.0
-
-[handlers]
-; NOTE: each key here must have a handler_xxx section below
-keys=console,overwrite_file
-
-[handler_console]
-class=StreamHandler
-level=DEBUG
-formatter=time_level_thread_name_message
-args=(sys.stdout,)
-
-[handler_overwrite_file]
-class=FileHandler
-level=DEBUG
-formatter=time_level_thread_name_message
-args=('pytest.log', 'w')
-
-[formatters]
-; NOTE: each key here must have a formatter_xxx section below
-keys=time_level_thread_name_message
-
-[formatter_time_level_thread_name_message]
-format=%(asctime)s [%(levelname)-1.1s] [%(threadName)-10.10s]: %(name)-14.14s - %(message)s
-; NOTE don't need milliseconds meanwhile, otherwise append .%03d below
-datefmt=%H:%M:%S
-
-[loggers]
-keys=root,selenium,urllib3,WDM
-
-[logger_root]
-level=DEBUG
-handlers=console,overwrite_file
-
-[logger_selenium]
-level=WARNING
-handlers=console,overwrite_file
-qualname=selenium
-propagate=0
-
-[logger_urllib3]
-level=DEBUG
-handlers=console,overwrite_file
-qualname=urllib3
-propagate=0
-
-[logger_WDM]
-level=WARNING
-handlers=console,overwrite_file
-qualname=WDM
-propagate=0
