+++ qa-testing-utils/MANIFEST.in
@@ -1,0 +1,1 @@
+include src/qa_testing_utils/logging.ini
